backend/requirements.txt
tigeropen>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
